academic_rag_system/requirements.txt
streamlit
langchain
langchain-community
langchain-core
qdrant-client
sentence-transformers
python-dotenv
pymupdf
requests
llama-cpp-python
latex2mathml>=3.73.0
Pillow>=9.0.0
numpy>=1.21.0